Make callisto buildable on Linux again and fix time-dependent bug
Note: Works for me on Ubuntu 20.04 (WSL) with Boost 1.81 and gcc/g++11.

Also fixes a (real?) bug that could cause callisto to wrongfully detect unexported resources in the ROM at specific times due to incorrectly discarded leading zeroes (don't ask).

+++ callisto/saver/marker.h
@@ -44,9 +44,9 @@ namespace callisto {
 
 		struct Extracted {
 			const uint16_t bitfield;
-			const uint64_t timestamp;
+			const long long timestamp;
 
-			Extracted(uint16_t bitfield, uint64_t timestamp) : bitfield(bitfield), timestamp(timestamp) {}
+			Extracted(uint16_t bitfield, long long timestamp) : bitfield(bitfield), timestamp(timestamp) {}
 		};
 
 		static uint16_t extractablesToBitField(const std::vector<ExtractableType>& extractables);
